+++ NEWS.md
@@ -13,7 +13,12 @@ NEW FEATURES :
 * `createClusterST()`/`editClusterST()` :  
   - **New properties** (*efficiencywithdrawal*, *penalize-variation-injection*, *penalize-variation-withdrawal*, see list of properties according to study version of Antares with `storage_values_default()`)  
   - **New optional time series** (cost-injection, cost-withdrawal, cost-level, cost-variation-injection, cost-variation-withdrawal)
+  - **New additional constraints** (properties and time series)
 * `updateScenarioBuilder()` New type of series "hfl" ("hydro final level", similar to "hydrolevels") is available
+
+
+NEW FEATURES (other) :  
+
 * `editBindingConstraint()` : control the dimensions of the matrix only if a time series is provided by the user for optimization
 * `.createCluster()` uses a specific endpoint to write cluster's metadata and commands to write matrix
 * Add new function `setThematicTrimming()` to set the thematic trimming in file `generaldata.ini`